refactor(llm-router): prefix model files with provider name
Rename gpt-5-2.ts → openai-gpt-5-2.ts and gpt-5-4.ts → openai-gpt-5-4.ts to follow a provider-prefixed naming convention for model files.
