+++ x/pierre/llm-router/src/clients/openai-responses/models/openai-gpt-5-2.ts
